+++ partials/mobile/mobile-icon.php
@@ -87,7 +87,7 @@ if ( has_nav_menu( $menu_location ) || $ms_global_menu ) :
 
 		<?php do_action( 'ocean_before_mobile_icon_inner' ); ?>
 
-		<a href="#" class="mobile-menu" aria-label="<?php esc_html_e( 'Mobile Menu', 'oceanwp' ); ?>">
+		<a href="#" class="mobile-menu" aria-label="<?php esc_attr_e( 'Mobile Menu', 'oceanwp' ); ?>">
 			<?php
 			if ( 'default' !== $btn ) {
 				?>