+++ sidebar-left.php
@@ -9,7 +9,7 @@
 
 <?php do_action( 'ocean_before_sidebar' ); ?>
 
-<aside id="left-sidebar" class="sidebar-container widget-area sidebar-secondary"<?php oceanwp_schema_markup( 'sidebar' ); ?> role="complementary" aria-label="<?php esc_html_e( 'Secondary Sidebar', 'oceanwp' ); ?>">
+<aside id="left-sidebar" class="sidebar-container widget-area sidebar-secondary"<?php oceanwp_schema_markup( 'sidebar' ); ?> role="complementary" aria-label="<?php esc_attr_e( 'Secondary Sidebar', 'oceanwp' ); ?>">
 
 	<?php do_action( 'ocean_before_sidebar_inner' ); ?>