+++ changelog.md
@@ -7,6 +7,9 @@
 * **Fixed** - Code optimized for speed improvements.
 * **Fixed** - Undefined variable $theme_version on menu settings page.
 * **Fixed** - Undefined variable $vh_width_minus when using vertical header.
+* **Fixed** - Woocommerce product quantity selector bug.
+* **Fixed** - Woocommerce product quantity selector bug on floating cart element.
+* **Fixed** - Woocommerce product quantity selector bug on qucik view.
 * **Fixed** - Customizer not loading - Uncaught TypeError: data.value.indexOf is not a function.
 * **Fixed** - Woocommerce 'Apply Coupon' button layout issue in mobile view.
 * **Fixed** - Translation issues - translation not reflected on the front end: search widget, related posts title, header search overlay, header replace search, next/previous post.